refactor(sdk): extract Sessions lifecycle to _client/_sessions.py
PR 6/N of the SDK refactor split (see #702 WebSocket, #706 Guardian,
#711 Discovery, #713 RFQ, #716 AI gateway for the mixin template).
Moves the A2A session lifecycle calls out of cullis_sdk/client.py
into a dedicated mixin under cullis_sdk/_client/.

What moves:

- cullis_sdk/_client/_sessions.py — new module exposing
  _SessionsMixin with the five lifecycle methods:
    open_session
    accept_session
    reject_session
    close_session
    list_sessions
  Plus the scoped import of SessionInfo from cullis_sdk.types.

- cullis_sdk/client.py:
  * CullisClient now inherits
    (_AiGatewayMixin, _DiscoveryMixin, _GuardianMixin, _RfqMixin,
     _SessionsMixin, _WebSocketMixin)
    in alphabetical order (A < D < G < R < S < W).
  * Five methods, their section header, and the now-unused
    SessionInfo import are removed.

- cullis_sdk/__init__.py — unchanged. SessionInfo continues to be
  re-exported via `from cullis_sdk.types import …`, same path as
  before.

No @deprecated annotations in this PR — the eventual deprecation of
the session lifecycle in favour of send_oneshot will land in a
dedicated PR once the client.py split is complete. This PR is
pure code movement, no behaviour change.

Routing logic is preserved byte-identically: proxy-bound clients
(from_connector / from_enrollment / from_identity_dir) keep going
through /v1/egress/sessions; direct-broker clients keep going
through /v1/broker/sessions. The reject-folds-into-close egress
quirk and the list-shape unwrap are kept exactly as they were.
