+++ cullis_sdk/client.py
@@ -38,12 +38,13 @@ from cullis_sdk.crypto.message_signer import (
     verify_signature,
 )
 from cullis_sdk.crypto.e2e import encrypt_for_agent, decrypt_from_agent
-from cullis_sdk.types import AgentInfo, SessionInfo, InboxMessage
+from cullis_sdk.types import AgentInfo, InboxMessage
 from cullis_sdk._logging import log, RED
 from cullis_sdk._client._ai_gateway import _AiGatewayMixin
 from cullis_sdk._client._discovery import PubkeyFetchError, _DiscoveryMixin
 from cullis_sdk._client._guardian import _GuardianMixin
 from cullis_sdk._client._rfq import _RfqMixin
+from cullis_sdk._client._sessions import _SessionsMixin
 from cullis_sdk._client._websocket import WebSocketConnection, _WebSocketMixin
 
 
@@ -223,7 +224,7 @@ def _check_insecure_tls(verify_tls: bool) -> None:
     )
 
 
-class CullisClient(_AiGatewayMixin, _DiscoveryMixin, _GuardianMixin, _RfqMixin, _WebSocketMixin):
+class CullisClient(_AiGatewayMixin, _DiscoveryMixin, _GuardianMixin, _RfqMixin, _SessionsMixin, _WebSocketMixin):
     """Client for the Cullis federated agent trust broker."""
 
     def __init__(
@@ -1666,99 +1667,6 @@ class CullisClient(_AiGatewayMixin, _DiscoveryMixin, _GuardianMixin, _RfqMixin, 
                 f"[{agent_id}] login_via_proxy_with_local_key failed "
                 f"(HTTP {e.response.status_code}): {e.response.text}"
             ) from e
-
-    # ── Sessions ────────────────────────────────────────────────────
-
-    def open_session(self, target_agent_id: str, target_org_id: str,
-                     capabilities: list[str]) -> str:
-        """Open a new session with a target agent. Returns session_id.
-
-        Proxy-bound clients (``from_connector``, ``from_enrollment``,
-        ``from_identity_dir``) route through ``/v1/egress/sessions`` —
-        the proxy's local mini-broker handles intra-org and falls
-        through to the broker bridge for cross-org. Direct-broker
-        clients keep using ``/v1/broker/sessions``.
-        """
-        if self._use_egress_for_sessions:
-            resp = self._egress_http("post", "/v1/egress/sessions", json={
-                "target_agent_id": target_agent_id,
-                "target_org_id": target_org_id,
-                "capabilities": capabilities,
-            })
-            resp.raise_for_status()
-            return resp.json()["session_id"]
-        path = "/v1/broker/sessions"
-        resp = self._authed_request("POST", path, json={
-            "target_agent_id": target_agent_id,
-            "target_org_id": target_org_id,
-            "requested_capabilities": capabilities,
-        })
-        resp.raise_for_status()
-        return resp.json()["session_id"]
-
-    def accept_session(self, session_id: str) -> None:
-        """Accept a pending session."""
-        if self._use_egress_for_sessions:
-            resp = self._egress_http(
-                "post", f"/v1/egress/sessions/{session_id}/accept"
-            )
-            resp.raise_for_status()
-            return
-        path = f"/v1/broker/sessions/{session_id}/accept"
-        resp = self._authed_request("POST", path)
-        resp.raise_for_status()
-
-    def reject_session(self, session_id: str) -> None:
-        """Reject a pending session.
-
-        The egress router has no dedicated reject endpoint — it folds
-        rejection into ``/close`` (the local store treats both as a
-        terminal state with the same semantics for the initiator).
-        Direct-broker clients still get the reject path so the broker
-        can distinguish 'rejected by target' from 'closed'.
-        """
-        if self._use_egress_for_sessions:
-            resp = self._egress_http(
-                "post", f"/v1/egress/sessions/{session_id}/close"
-            )
-            resp.raise_for_status()
-            return
-        path = f"/v1/broker/sessions/{session_id}/reject"
-        resp = self._authed_request("POST", path)
-        resp.raise_for_status()
-
-    def close_session(self, session_id: str) -> None:
-        """Close an active session."""
-        if self._use_egress_for_sessions:
-            resp = self._egress_http(
-                "post", f"/v1/egress/sessions/{session_id}/close"
-            )
-            resp.raise_for_status()
-            return
-        path = f"/v1/broker/sessions/{session_id}/close"
-        resp = self._authed_request("POST", path)
-        resp.raise_for_status()
-
-    def list_sessions(self, status: str | None = None) -> list[SessionInfo]:
-        """List sessions, optionally filtered by status.
-
-        The egress shape wraps the list in ``{"sessions": [...]}`` and
-        the broker shape returns a flat list — unwrap on the egress
-        side so callers see the same return type.
-        """
-        params = {}
-        if status:
-            params["status"] = status
-        if self._use_egress_for_sessions:
-            resp = self._egress_http("get", "/v1/egress/sessions", params=params)
-            resp.raise_for_status()
-            body = resp.json()
-            sessions = body.get("sessions", []) if isinstance(body, dict) else body
-            return [SessionInfo.from_dict(s) for s in sessions]
-        path = "/v1/broker/sessions"
-        resp = self._authed_request("GET", path, params=params)
-        resp.raise_for_status()
-        return [SessionInfo.from_dict(s) for s in resp.json()]
 
     # ── Messaging ───────────────────────────────────────────────────
